# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "save",
              "approve",
              "reject"
            ],
            "title": "Action",
            "type": "string"
          },
          "collection_id": {
            "minLength": 1,
            "title": "Collection Id",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "question": {
            "maxLength": 2000,
            "title": "Question",
            "type": "string"
          },
          "reference_answer": {
            "maxLength": 10000,
            "title": "Reference Answer",
            "type": "string"
          },
          "support_positions": {
            "items": {
              "type": "integer"
            },
            "maxItems": 1000,
            "title": "Support Positions",
            "type": "array"
          }
        },
        "required": [
          "expected_revision",
          "collection_id",
          "question",
          "reference_answer",
          "support_positions",
          "action"
        ],
        "title": "CandidateReviewRequest",
        "type": "object"
      },
...
      "GeneratedCandidate": {
        "properties": {
          "collection_id": {
            "title": "Collection Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "multi_chunk": {
            "title": "Multi Chunk",
            "type": "boolean"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reference_answer": {
            "title": "Reference Answer",
            "type": "string"
          },
          "reference_chunks": {
            "items": {
              "$ref": "#/components/schemas/CandidateChunk"
            },
            "title": "Reference Chunks",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "slot_index": {
            "title": "Slot Index",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "support_positions": {
            "items": {
              "type": "integer"
            },
            "title": "Support Positions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "run_id",
          "collection_id",
          "revision",
          "slot_index",
          "question",
          "reference_answer",
          "reference_chunks",
          "support_positions",
          "multi_chunk",
          "status",
          "prompt_version",
          "model_name",
          "created_at"
        ],
        "title": "GeneratedCandidate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportFailure": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ImportIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "error",
          "issues"
        ],
        "title": "ImportFailure",
        "type": "object"
      },
...
      "PublicationRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Candidate Ids",
            "type": "array"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "dataset_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "required": [
          "candidate_ids"
        ],
        "title": "PublicationRequest",
        "type": "object"
      },
...
      "VersionSummary": {
        "properties": {
          "case_count": {
            "title": "Case Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_collection_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Collection Ids",
            "type": "array"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "version",
          "case_count",
          "source_filename",
          "created_at"
        ],
        "title": "VersionSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ragEva API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/candidates/publish": {
      "post": {
        "operationId": "publish_candidates_api_v1_candidates_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFailure"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Publish Candidates",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}": {
      "get": {
        "operationId": "get_candidate_api_v1_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedCandidate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "generations"
        ]
      },
      "patch": {
        "operationId": "review_candidate_api_v1_candidates__candidate_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedCandidate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewFailure"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Review Candidate",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/duplicate-check": {
      "get": {
        "operationId": "get_duplicate_check_api_v1_candidates__candidate_id__duplicate_check_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DuplicateCheck"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Duplicate Check Api V1 Candidates  Candidate Id  Duplicate Check Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Duplicate Check",
        "tags": [
          "generations"
        ]
      },
      "post": {
        "operationId": "recheck_duplicate_api_v1_candidates__candidate_id__duplicate_check_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recheck Duplicate",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/duplicate-decision": {
      "post": {
        "operationId": "allow_suspected_api_v1_candidates__candidate_id__duplicate_decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Allow Suspected",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/duplicate-history": {
      "get": {
        "operationId": "get_duplicate_history_api_v1_candidates__candidate_id__duplicate_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DuplicateCheck"
                  },
                  "title": "Response Get Duplicate History Api V1 Candidates  Candidate Id  Duplicate History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Duplicate History",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/candidates/{candidate_id}/revisions": {
      "get": {
        "operationId": "get_candidate_revisions_api_v1_candidates__candidate_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CandidateRevision"
                  },
                  "title": "Response Get Candidate Revisions Api V1 Candidates  Candidate Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate Revisions",
        "tags": [
          "generations"
        ]
      }
    },
...
    "/api/v1/document-collections/{collection_id}/generations": {
      "post": {
        "operationId": "create_api_v1_document_collections__collection_id__generations_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "generations"
        ]
      }
    },
...
    "/api/v1/generations/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_generations__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/generations/{run_id}/candidates": {
      "get": {
        "operationId": "get_candidates_api_v1_generations__run_id__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedCandidate"
                  },
                  "title": "Response Get Candidates Api V1 Generations  Run Id  Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidates",
        "tags": [
          "generations"
        ]
      }
    },
    "/api/v1/generations/{run_id}/usage": {
      "get": {
        "operationId": "get_usage_api_v1_generations__run_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Usage",
        "tags": [
          "generations"
        ]
      }
    },
...
  }
}